LPS-148877 Move package to portlet.display.template (commerce-order-content-web)
